Move App component to components dir
